requirements.txt
pip
autopep8
Flask
gunicorn==22.0.0
Werkzeug
Flask-Bcrypt
python-dotenv
flask-jwt-extended
flask-sqlalchemy 
google-generativeai 
